Update for international keyboard.
Framework offers keyboards with a #/~ key to the left of the enter, and
\/| to the left of Z; and ones with no #/~, and \/| above enter. I
prefer the second kind, but it means I need to put # above 3 and ~ above
`. I lose £ and ¬ but XCompose has both of those. I could keep the
current layout on kinesis but that seems confusing.

Fortunately both my work and personal laptops are now framework, so I
don't need an extra layout switch.